7.1.1 reorganisations folders, & create modal template
