--- src/Readme.md
@@ -1,1 +1,0 @@
-Папка для скриптов.
