+++ src/model/comments.js
@@ -1,0 +1,8 @@
+import Abstract from '../utils/abstract/abstract.js';
+
+export default class Comments extends Abstract {
+  constructor() {
+    super();
+  }
+
+}
