+++ src/model/movies.js
@@ -1,0 +1,8 @@
+import Abstract from '../utils/abstract/abstract.js';
+
+export default class Movies extends Abstract {
+  constructor() {
+    super();
+  }
+
+}
