+++ src/utils/abstract/smart.js
@@ -1,7 +1,7 @@
 /*
 * обновляет данные и, если нужно, вызывает перерисовку
 * */
-import Abstract from './abstract';
+import Abstract from './abstract.js';
 
 export default class Smart extends Abstract {
   constructor() {
