+++ src/view/film-board.js
@@ -1,4 +1,4 @@
-import Abstract from './abstract.js';
+import Abstract from '../utils/abstract/abstract.js';
 
 const createFilmsBoardTemplate = () => (
   '<section class="films"></section>'
