+++ src/view/show-more-btn.js
@@ -1,4 +1,4 @@
-import Abstract from './abstract.js';
+import Abstract from '../utils/abstract/abstract.js';
 
 export const createShowMoreTemplate = () => (
   '<button class="films-list__show-more">Show more</button>'
